extracted out function getMarketingDataForQuestion function seperate helper file
